FSAL_LUSTRE/Upcalls: this commit implements FSAL Upcalls for FSAL_LUSTRE
This commit implements the basic support of upcalls in FSAL_LUSTRE.
The mechanism is based on Lustre changelogs feature. For scalability
reason, changelog are accessed through the lcap_* API intsead of the
llapi_* (lcap provides changelog caching which leverage loads on the
Lustre MDS). Upcalls are generated by a thread, just the way GPFS does
and uses fsal_up_top function vector.
This commit needs to have lustre support for jobid in changelog records.
This means that LUSTRE 2.7 or higher is required to compile this commit.
The feature is triggered by cmake and/or rpmbuild. By default it is OFF.

Change-Id: I5f278e3f371fa0bbe2d227fd5f18bc14be0f239b
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>