Config_parsing : add the optional support en environment variables in the configuration file.
This commits adds the support of environment variables in the configuration file.
By default, the feature is disabled, it is activated by cmake.
The feature allows to have left values defined by a variable set in the environment in configuration stanzas.
This is pretty useful to keep a single configuration file with various testing environment in a testbed.

Change-Id: I79b8e54e9e53a9169c9ae791a60693a88224502b
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>